Rework which-key--get-current-bindings
Getting the bindings for a a prefix by parsing the buffer produced by
`describe-buffer-bindings` seems a little convoluted. Especially when
`map-keymap` seems to be just the job and has the added advantage of
giving us the actual values of the bindings in the current keymap. This
doesn't matter too much for a bound command, but when the binding is to
a keymap it lets us attempt to get `(keymap-prompt binding)` as a
potential name.

I'd not be surprised if it's faster too since we're not mucking about
with regular expressions.