Uniquify keys in -get-raw-current-bindings
Yay for free software! After taking a look at the implementation of
`map-keymap` it's apparent that it respects the ordering of a keymap so
the first entry we get for a given key will be the binding that would be
triggered if we were to actually type that key sequence. So
`-get-raw-current-bindings` now does an assoc look up on the key
description and skips any that have already been seen.

Fixes the issue with C-x u showing up twice while `global-undo-tree` is
turned on (and other, similar cases).