Add support for Devil
What is Devil?

> Devil mode trades your comma key in exchange for a modifier-free
editing experience in Emacs.

The above description comes from their README at:

https://github.com/susam/devil

Unfortunately, Devil is not currently supported by which-key. For
example, when I type `, x`, the which-key popup for `C-x` does not
appear. Instead, I merely see the following text in the echo area:

```
Devil: C-x
```

With this patch, I have mimicked the existing god-mode logic in
which-key to intercept the current Devil prompt and spawn a which-key
popup.

This feature was requested by susam/devil#3:

- https://github.com/susam/devil/issues/3