Restore window scroll after popup close
Currently, if displaying the popup would hide point, the window is
scrolled to prevent this; but it is not scrolled back when the popup
closes. This can be disorienting when the popup is tall, as the scroll
changes by a large amount.

The solution is to store the scroll position of the selected window
before showing the popup, and restore it after closing it.