Symbolize special keys
Use symbols to represent special keys. If value of
``which-key-special-keys`` is an *alist* of cons
cells, whose cars are special keys such as SPC,
TAB,  ... and the cdrs are symbol strings, replace
the special keys with corresponding symbols.

For backward-compatibility, if
``which-key-special-keys`` is an ordinary list of
keys, old logic of trucation to single character
works.